+++ src/frequenz/sdk/timeseries/consumer.py
@@ -38,8 +38,7 @@ class Consumer:
         from frequenz.sdk.timeseries import ResamplerConfig
 
         await microgrid.initialize(
-            "127.0.0.1",
-            50051,
+            "grpc://127.0.0.1:50051",
             ResamplerConfig(resampling_period=timedelta(seconds=1.0))
         )