+++ src/frequenz/sdk/timeseries/grid.py
@@ -51,8 +51,7 @@ class Grid:
         from frequenz.sdk.timeseries import ResamplerConfig
 
         await microgrid.initialize(
-            "127.0.0.1",
-            50051,
+            "grpc://127.0.0.1:50051",
             ResamplerConfig(resampling_period=timedelta(seconds=1))
         )