+++ tests/actor/test_data_sourcing.py
@@ -40,7 +40,7 @@ async def test_data_sourcing_actor() -> None:
     servicer.add_connection(7, 8)
     servicer.add_connection(8, 9)
 
-    await connection_manager.initialize("[::1]", 57899)
+    await connection_manager.initialize("grpc://[::1]:57899")
 
     req_chan = Broadcast[ComponentMetricRequest](name="data_sourcing_requests")
     req_sender = req_chan.new_sender()